Ora nel grafico ci sono anche i numeri di fianco ad ogni pallino. Comandi di gnuplot nel relativo File. Comandi del parser nel relativo File
