+++ PRO2/coordinateAtt48.txt
@@ -1,48 +1,48 @@
-6734.000000 1453.000000 
-2233.000000 10.000000 
-5530.000000 1424.000000 
-401.000000 841.000000 
-3082.000000 1644.000000 
-7608.000000 4458.000000 
-7573.000000 3716.000000 
-7265.000000 1268.000000 
-6898.000000 1885.000000 
-1112.000000 2049.000000 
-5468.000000 2606.000000 
-5989.000000 2873.000000 
-4706.000000 2674.000000 
-4612.000000 2035.000000 
-6347.000000 2683.000000 
-6107.000000 669.000000 
-7611.000000 5184.000000 
-7462.000000 3590.000000 
-7732.000000 4723.000000 
-5900.000000 3561.000000 
-4483.000000 3369.000000 
-6101.000000 1110.000000 
-5199.000000 2182.000000 
-1633.000000 2809.000000 
-4307.000000 2322.000000 
-675.000000 1006.000000 
-7555.000000 4819.000000 
-7541.000000 3981.000000 
-3177.000000 756.000000 
-7352.000000 4506.000000 
-7545.000000 2801.000000 
-3245.000000 3305.000000 
-6426.000000 3173.000000 
-4608.000000 1198.000000 
-23.000000 2216.000000 
-7248.000000 3779.000000 
-7762.000000 4595.000000 
-7392.000000 2244.000000 
-3484.000000 2829.000000 
-6271.000000 2135.000000 
-4985.000000 140.000000 
-1916.000000 1569.000000 
-7280.000000 4899.000000 
-7509.000000 3239.000000 
-10.000000 2676.000000 
-6807.000000 2993.000000 
-5185.000000 3258.000000 
-3023.000000 1942.000000 
+6734.000000 1453.000000 1 
+2233.000000 10.000000 2 
+5530.000000 1424.000000 3 
+401.000000 841.000000 4 
+3082.000000 1644.000000 5 
+7608.000000 4458.000000 6 
+7573.000000 3716.000000 7 
+7265.000000 1268.000000 8 
+6898.000000 1885.000000 9 
+1112.000000 2049.000000 10 
+5468.000000 2606.000000 11 
+5989.000000 2873.000000 12 
+4706.000000 2674.000000 13 
+4612.000000 2035.000000 14 
+6347.000000 2683.000000 15 
+6107.000000 669.000000 16 
+7611.000000 5184.000000 17 
+7462.000000 3590.000000 18 
+7732.000000 4723.000000 19 
+5900.000000 3561.000000 20 
+4483.000000 3369.000000 21 
+6101.000000 1110.000000 22 
+5199.000000 2182.000000 23 
+1633.000000 2809.000000 24 
+4307.000000 2322.000000 25 
+675.000000 1006.000000 26 
+7555.000000 4819.000000 27 
+7541.000000 3981.000000 28 
+3177.000000 756.000000 29 
+7352.000000 4506.000000 30 
+7545.000000 2801.000000 31 
+3245.000000 3305.000000 32 
+6426.000000 3173.000000 33 
+4608.000000 1198.000000 34 
+23.000000 2216.000000 35 
+7248.000000 3779.000000 36 
+7762.000000 4595.000000 37 
+7392.000000 2244.000000 38 
+3484.000000 2829.000000 39 
+6271.000000 2135.000000 40 
+4985.000000 140.000000 41 
+1916.000000 1569.000000 42 
+7280.000000 4899.000000 43 
+7509.000000 3239.000000 44 
+10.000000 2676.000000 45 
+6807.000000 2993.000000 46 
+5185.000000 3258.000000 47 
+3023.000000 1942.000000 48
